smt-aws-apigateway-parent/smt-aws-lambda-soap-stub-parent/smt-aws-lambda-soap-stub-api/src/main/java/shiver/me/timbers/aws/lambda/soap/stub/VerifyFailure.java
package shiver.me.timbers.aws.lambda.soap.stub;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

@JsonIgnoreProperties(ignoreUnknown = true)
public class VerifyFailure {

    private String message;

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

    public String getErrorMessage() {
        return message;
    }

    public void setErrorMessage(String message) {
        this.message = message;
    }
}
